Improved cellpose prediction script

+++ scripts/segmentation/cellpose_downscaling.py
@@ -16,9 +16,11 @@ import pandas as pd
 # Common variables
 load_dotenv(".env")
 DATA_ROOT = Path(os.environ.get("DATA_ROOT"))
-dir_labeled = DATA_ROOT / "processed/labeled/Au_01-vol_01/labeled-01"
-dir_output = DATA_ROOT / "processed/segmented/Au_01-vol_01/downscaling_tests"
-dir_output.mkdir(parents=True, exist_ok=True)
+DIR_LABELED = DATA_ROOT / "processed/labeled/Au_01-vol_01/labeled-01"
+DIR_OUTPUT = DATA_ROOT / "processed/segmented/Au_01-vol_01/downscaling_tests"
+DIR_OUTPUT.mkdir(parents=True, exist_ok=True)
+DOWNSCALE_FACTORS = [1, 2, 4, 5, 8, 10, 16, 20, 32]
+SEG_TARGET = "cells"
 
 
 # Function definitions
@@ -35,31 +37,36 @@ def imwrite(fname: Path, image: np.ndarray) -> None:
 
 
 def main() -> None:
-    image_path = dir_labeled / "Au_01-vol_01-z_0425.tif"
-    factors = [1, 2, 4, 5, 8, 10, 16, 20, 32]
     timing_results = []
 
-    for downscale_factor in factors:
-        image = imread_downscaled(image_path, downscale_factor)
+    for image_path in DIR_LABELED.glob("*.tif"):
+        mask_path = DIR_LABELED / f"labels/{image_path.stem}_{SEG_TARGET}.tif"
+        if not mask_path.exists():
+            continue
 
-        cellpose_model = CellposeModel(gpu=True)
+        for dowscale_factor in DOWNSCALE_FACTORS:
+            image = imread_downscaled(image_path, dowscale_factor)
 
-        start_time = time.time()
-        output = cellpose_model.eval(image)
-        elapsed_time = time.time() - start_time
+            cellpose_model = CellposeModel(gpu=True)
 
-        # Store downsample factor and corresponding time
-        timing_results.append((downscale_factor, elapsed_time))
-        print(f"Downscale factor {downscale_factor}: {elapsed_time:.2f} seconds")
+            start_time = time.time()
+            output = cellpose_model.eval(image)
+            elapsed_time = time.time() - start_time
 
-        imwrite(dir_output / f"{image_path.stem}-ds{downscale_factor}.tif", image)
-        imwrite(
-            dir_output / f"cellposeSAM-pretrained-mask-ds{downscale_factor}.tif",
-            output[0],
-        )
+            # Store downsample factor and corresponding time
+            timing_results.append((dowscale_factor, elapsed_time))
+            print(f"Downscale factor {dowscale_factor}: {elapsed_time:.2f} seconds")
+
+            imwrite(
+                DIR_OUTPUT
+                / f"{image_path.stem}-cellposeSAM-pretrained-{SEG_TARGET}-ds{dowscale_factor}.tif",
+                output[0],
+            )
 
     timing_df = pd.DataFrame(timing_results, columns=["downscale_factor", "time"])
-    timing_df.to_csv(dir_output / "cellposeSAM-downscaled-time.csv", index=False)
+    timing_df.to_csv(
+        DIR_OUTPUT / f"cellposeSAM-{SEG_TARGET}-downscaled-time.csv", index=False
+    )
 
 
 if __name__ == "__main__":
